chore: move custom lxplus custom exceptions to a dedicated script
