+++ etl/python/common/lxplus_exceptions.py
@@ -1,0 +1,18 @@
+class SSHAuthenticationTimeoutError(Exception):
+    pass
+
+
+class SSHAuthenticationFailedError(Exception):
+    pass
+
+
+class XrdcpNoServersAvailableToReadFileError(Exception):
+    pass
+
+
+class XrdcpTimeoutError(Exception):
+    pass
+
+
+class XrdcpUnknownError(Exception):
+    pass
